+++ doc/changelog.rst
@@ -7,6 +7,7 @@ Version 0.7
 * Add a new ``pybench`` program, similar to ``python3 -m perf``
 * Add a new ``dump`` command
 * ``convert`` command: add ``--indent`` and ``--remove-warmups`` options
+* New :func:`add_runs` function.
 
 Version 0.6 (2016-07-06)
 ------------------------